+++ platformplugin/dframewindow.h
@@ -80,6 +80,8 @@ public:
     void disableRepaintShadow();
     void enableRepaintShadow();
 
+    bool redirectContent() const;
+
 signals:
     void contentMarginsHintChanged(const QMargins &oldMargins) const;
 
@@ -107,7 +109,7 @@ private:
 
     void updateShadow();
     void updateShadowAsync(int delaye = 30);
-    void updateContentMarginsHint();
+    void updateContentMarginsHint(bool force = false);
     void updateMask();
     void updateFrameMask();
 
@@ -117,6 +119,8 @@ private:
     void startCursorAnimation();
 
     bool disableFrame() const;
+
+    void onDevicePixelRatioChanged();
 
     static QList<DFrameWindow*> frameWindowList;