+++ platformplugin/dplatformnativeinterfacehook.cpp
@@ -46,7 +46,7 @@ static DeviceType _inputEventSourceDevice(const QInputEvent *event)
 }
 #endif
 
-QFunctionPointer DPlatformNativeInterfaceHook::platformFunction(QPlatformNativeInterface *interface, const QByteArray &function)
+static QFunctionPointer getFunction(const QByteArray &function)
 {
     if (function == setWmBlurWindowBackgroundArea) {
         return reinterpret_cast<QFunctionPointer>(&Utility::blurWindowBackground);
@@ -55,17 +55,23 @@ QFunctionPointer DPlatformNativeInterfaceHook::platformFunction(QPlatformNativeI
     } else if (function == setWmBlurWindowBackgroundMaskImage) {
         return reinterpret_cast<QFunctionPointer>(&Utility::blurWindowBackgroundByImage);
     } else if (function == hasBlurWindow) {
-        return reinterpret_cast<QFunctionPointer>(&Utility::hasBlurWindow);
+        return reinterpret_cast<QFunctionPointer>(&DWMSupport::Global::hasBlurWindow);
     } else if (function == hasComposite) {
-        return reinterpret_cast<QFunctionPointer>(&Utility::hasComposite);
+        return reinterpret_cast<QFunctionPointer>(&DWMSupport::Global::hasComposite);
+    } else if (function == hasNoTitlebar) {
+        return reinterpret_cast<QFunctionPointer>(&DWMSupport::Global::hasNoTitlebar);
+    } else if (function == hasWindowAlpha) {
+        return reinterpret_cast<QFunctionPointer>(&DWMSupport::Global::hasWindowAlpha);
     } else if (function == windowManagerName) {
-        return reinterpret_cast<QFunctionPointer>(&Utility::windowManagerName);
+        return reinterpret_cast<QFunctionPointer>(&DWMSupport::Global::windowManagerName);
     } else if (function == connectWindowManagerChangedSignal) {
         return reinterpret_cast<QFunctionPointer>(&DWMSupport::connectWindowManagerChangedSignal);
     } else if (function == connectHasBlurWindowChanged) {
         return reinterpret_cast<QFunctionPointer>(&DWMSupport::connectHasBlurWindowChanged);
     } else if (function == connectHasCompositeChanged) {
         return reinterpret_cast<QFunctionPointer>(&DWMSupport::connectHasCompositeChanged);
+    } else if (function == connectHasNoTitlebarChanged) {
+        return reinterpret_cast<QFunctionPointer>(&DWMSupport::connectHasNoTitlebarChanged);
     } else if (function == getWindows) {
         return reinterpret_cast<QFunctionPointer>(&Utility::getWindows);
     } else if (function == getCurrentWorkspaceWindows) {
@@ -85,7 +91,7 @@ QFunctionPointer DPlatformNativeInterfaceHook::platformFunction(QPlatformNativeI
     } else if (function == popupSystemWindowMenu) {
         return reinterpret_cast<QFunctionPointer>(&DWMSupport::popupSystemWindowMenu);
     } else if (function == setWindowProperty) {
-        return reinterpret_cast<QFunctionPointer>(&DPlatformWindowHelper::setWindowProperty);
+        return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::setWindowProperty);
     } else if (function == pluginVersion) {
         return reinterpret_cast<QFunctionPointer>(&version);
     } else if (function == inputEventSourceDevice) {
@@ -102,6 +108,31 @@ QFunctionPointer DPlatformNativeInterfaceHook::platformFunction(QPlatformNativeI
         return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::enableDxcb);
     } else if (function == isEnableDxcb) {
         return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::isEnableDxcb);
+    } else if (function == setEnableNoTitlebar) {
+        return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::setEnableNoTitlebar);
+    } else if (function == isEnableNoTitlebar) {
+        return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::isEnableNoTitlebar);
+    } else if (function == buildNativeSettings) {
+        return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::buildNativeSettings);
+    } else if (function == clearNativeSettings) {
+        return reinterpret_cast<QFunctionPointer>(&DPlatformIntegration::clearNativeSettings);
+    }
+
+    return nullptr;
+}
+
+thread_local QHash<QByteArray, QFunctionPointer> DPlatformNativeInterfaceHook::functionCache;
+QFunctionPointer DPlatformNativeInterfaceHook::platformFunction(QPlatformNativeInterface *interface, const QByteArray &function)
+{
+    if (QFunctionPointer f = functionCache.value(function)) {
+        return f;
+    }
+
+    QFunctionPointer f = getFunction(function);
+
+    if (f) {
+        functionCache.insert(function, f);
+        return f;
     }
 
 #if QT_VERSION >= QT_VERSION_CHECK(5, 4, 0)